+++ pkg/epp/datalayer/metrics/datasource_test.go
@@ -25,11 +25,11 @@ import (
 	"k8s.io/apimachinery/pkg/types"
 
 	"sigs.k8s.io/gateway-api-inference-extension/pkg/epp/datalayer"
-	httpds "sigs.k8s.io/gateway-api-inference-extension/pkg/epp/datalayer/http"
+	"sigs.k8s.io/gateway-api-inference-extension/pkg/epp/datalayer/http"
 )
 
 func TestDatasource(t *testing.T) {
-	source := httpds.NewHTTPDataSource("https", "/metrics", true, MetricsDataSourceType,
+	source := http.NewHTTPDataSource("https", "/metrics", true, MetricsDataSourceType,
 		"metrics-data-source", parseMetrics, PrometheusMetricType)
 	extractor, err := NewModelServerExtractor(defaultTotalQueuedRequestsMetric, "", "", "", "")
 	assert.Nil(t, err, "failed to create extractor")
